+++ testtinymce/testapp/admin.py
@@ -1,4 +1,3 @@
-from django import forms
 from django.core.urlresolvers import reverse
 from django.contrib import admin
 from django.contrib.flatpages.admin import FlatPageAdmin
@@ -10,7 +9,7 @@ from tinymce.widgets import TinyMCE
 class TinyMCEFlatPageAdmin(FlatPageAdmin):
     def formfield_for_dbfield(self, db_field, **kwargs):
         if db_field.name == 'content':
-            return forms.CharField(widget=TinyMCE(
+            return db_field.formfield(widget=TinyMCE(
                 attrs={'cols': 80, 'rows': 30},
                 mce_attrs={'external_link_list_url': reverse('tinymce.views.flatpages_link_list')},
             ))
@@ -19,7 +18,7 @@ class TinyMCEFlatPageAdmin(FlatPageAdmin):
 class TinyMCETestPageAdmin(admin.ModelAdmin):
     def formfield_for_dbfield(self, db_field, **kwargs):
         if db_field.name in ('content1', 'content2'):
-            return forms.CharField(widget=TinyMCE(
+            return db_field.formfield(widget=TinyMCE(
                 attrs={'cols': 80, 'rows': 30},
                 mce_attrs={'external_link_list_url': reverse('tinymce.views.flatpages_link_list')},
             ))